+++ app/src/main/java/com/example/fitfluent/data/User.kt
@@ -6,7 +6,7 @@ data class User(
     var height_in_cm: Int,
     var weight_in_kg: Int,
     var age: Int,
-    var calorie_intake: Int,
+    var calorie_intake: Double,
     var calorie_time: String,
     var activity_level: Int
 ) : java.io.Serializable{
